backend/requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
langchain
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers
faiss-cpu
SQLAlchemy
pydantic
pydantic-settings
python-dotenv
python-multipart
